+++ connections/mongodbconn.py
@@ -1,22 +1,13 @@
 from pymongo import MongoClient
 from gridfs import GridFS
 from bson.objectid import ObjectId
-from os import environ, path
-from json import load
+from os import environ
+from ..settings import json_settings
 
-client = None
-
-try:
-    if client == None:
-        settings = path.abspath(path.join(path.dirname( __file__ ),'..','settings.json'))
-        with open(settings) as f:
-            json_settings = load(f)
-            if environ["analyzer_env"] == "local":
-                client = MongoClient(json_settings["mongo_settings_local"])
-            elif environ["analyzer_env"] == "docker":
-                client = MongoClient(json_settings["mongo_settings_docker"])
-except:
-    client = None
+if environ["analyzer_env"] == "local":
+    client = MongoClient(json_settings["mongo_settings_local"])
+elif environ["analyzer_env"] == "docker":
+    client = MongoClient(json_settings["mongo_settings_docker"])
 
 def update_item(db,col,_id,_set):
     item = client[db][col].find_one_and_update({'_id': _id},{'$set': _set})
@@ -52,13 +43,6 @@ def find_items(db,_set):
         return _list
     else:
         return ""
-
-#def add_item_fs(db,filebuffer,name,_set,uuid,_type):
-#    item = GridFS(client[db]).put(filebuffer,filename=name,metadata=_set,uuid=uuid,type=_type,encoding='utf-8')
-#    if item != None:
-#        return item
-#    else:
-#        return False
 
 def add_item_fs(db, col,filebuffer,name,_set,uuid,_type,time):
     item = GridFS(client[db]).put(filebuffer,filename=name,uuid=uuid,content_type=_type,encoding='utf-8')