+++ settings.py
@@ -1,0 +1,6 @@
+json_settings = {
+	"mongo_settings_local":"mongodb://localhost:27017/",
+	"mongo_settings_docker":"mongodb://mongodb:27017/",
+	"function_timeout":120,
+	"analyzer_timeout":130
+}